Lab 4

+++ database/migrations/2025_10_25_094437_drop_category_column_from_products_table.php
@@ -8,7 +8,9 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('category');
+            if (Schema::hasColumn('products', 'category')) {
+                $table->dropColumn('category');
+            }
         });
     }
